Alteracao no tipo da coluna nota da entidade AvaliacaoArtista para double

+++ ScreenSound.API/Requests/AvaliacaoArtistaRequest.cs
@@ -2,4 +2,4 @@ using System;
 
 namespace ScreenSound.API.Requests;
 
-public record AvaliacaoArtistaRequest(int ArtistaId, int Nota);
+public record AvaliacaoArtistaRequest(int ArtistaId, double Nota);
